# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.Container:
    properties:
      auto_remove:
        type: boolean
      cap_add:
        items:
          type: string
        type: array
      cap_drop:
        items:
          type: string
        type: array
      command:
        maxLength: 1024
        type: string
      container_id:
        type: string
      created_at:
        type: string
      dead:
        type: boolean
      detailed_network_info:
        additionalProperties:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.DetailedNetwork'
        type: object
      detailed_volume_info:
        additionalProperties:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.DetailedVolume'
        type: object
      entrypoint:
        description: Fields from inspect conversion (might be redundant)
        items:
          type: string
        type: array
      env_vars:
        items:
          type: string
        type: array
      exit_code:
        type: integer
      exposed_ports:
        items:
          type: string
        type: array
      finished_at:
        type: string
      has_changed:
        type: boolean
      health:
        type: string
      health_log:
        type: string
      healthcheck:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      host_ipc:
        type: boolean
      host_pid:
        type: boolean
      id:
        type: integer
      image:
        type: string
      image_id:
        type: string
      ip_address:
        type: string
      labels:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      last_inspected:
        type: string
      mounts:
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MountPoint'
        type: array
      name:
        maxLength: 255
        type: string
      names:
        description: Fields from list item conversion (might be redundant with inspect)
        items:
          type: string
        type: array
      network_mode:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkMode'
      networks:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      notes:
        maxLength: 5000
        type: string
      oom_killed:
        type: boolean
      paused:
        type: boolean
      platform:
        type: string
      ports:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      privileged:
        type: boolean
      processes:
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.Process'
        type: array
      read_only:
        type: boolean
      resource_limits:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ResourceLimits'
      resources:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      restart_count:
        type: integer
      restart_policy:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.RestartPolicy'
      restarting:
        type: boolean
      running:
        type: boolean
      secrets:
        items:
          type: string
        type: array
      security_info:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SecurityInfo'
      security_options:
        items:
          type: string
        type: array
      security_profile:
        type: string
      size_root_fs:
        type: integer
      size_rw:
        type: integer
      started_at:
        type: string
      state:
        description: Raw state string
        type: string
      stats:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerStats'
      status:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerStatus'
      up_time:
        type: string
      updated_at:
        type: string
      user:
        type: string
      user_id:
        type: integer
      userns_mode:
        type: string
      volumes:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      working_dir:
        type: string
    required:
    - image_id
    - name
    - network_mode
    - restart_policy
    - user_id
    type: object
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerCreateRequest:
    description: Configuration details for creating a new Docker container.
    properties:
      auto_remove:
        description: |-
          AutoRemove automatically removes the container when it exits.
          example: false
        example: false
        type: boolean
      command:
        description: |-
          Command is the command to run when the container starts.
          example: ["nginx", "-g", "daemon off;"]
        example:
        - nginx
        - -g
        - daemon off;
        items:
          type: string
        type: array
      cpu_limit:
        description: |-
          CPULimit specifies the CPU quota for the container (e.g., 1.5 means 1.5 CPU cores). Docker uses NanoCPUs (1e9 per core).
          example: 1.5
        example: 1.5
        type: number
      entrypoint:
        description: |-
          Entrypoint overrides the default entrypoint of the image.
          example: ["/docker-entrypoint.sh"]
        example:
        - /docker-entrypoint.sh
        items:
          type: string
        type: array
      env:
        description: |-
          Env is a list of environment variables to set in the container (e.g., "VAR=value").
          example: ["NGINX_HOST=example.com", "NGINX_PORT=80"]
        example:
        - NGINX_HOST=example.com
        - NGINX_PORT=80
        items:
          type: string
        type: array
      image:
        description: |-
          Image is the name or ID of the Docker image to use.
          required: true
          example: "nginx:latest"
        example: nginx:latest
        type: string
      labels:
        additionalProperties:
          type: string
        description: |-
          Labels are key-value pairs to apply to the container.
          example: {"environment": "development", "app": "webserver"}
        example:
          app: webserver
          environment: development
        type: object
      memory_limit:
        description: |-
          MemoryLimit is the maximum amount of memory the container can use (in bytes).
          example: 104857600 (100MB)
        example: 104857600
        type: integer
      name:
        description: |-
          Name is the name to assign to the new container.
          required: true
          example: "my-nginx-container"
        example: my-nginx-container
        type: string
      networks:
        description: |-
          Networks is a list of network names or IDs to connect the container to.
          example: ["my-app-network"]
        example:
        - my-app-network
        items:
          type: string
        type: array
      notes:
        description: |-
          Notes are user-defined notes for the container (stored in application DB).
          example: "Main web server container."
        example: Main web server container.
        type: string
      ports:
        description: Ports specifies the port mappings between the host and the container.
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.PortMapping'
        type: array
      privileged:
        description: |-
          Privileged gives the container extended privileges on the host. Use with caution.
          example: false
        example: false
        type: boolean
      restart_policy:
        description: |-
          RestartPolicy defines the container's behavior when it exits.
          example: "unless-stopped"
        enum:
        - "no"
        - on-failure
        - always
        - unless-stopped
        example: unless-stopped
        type: string
      volumes:
        description: Volumes specifies the volume mappings between the host/named
          volumes and the container.
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeMapping'
        type: array
    required:
    - image
    - name
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerListResponse:
    description: Contains a list of container details along with pagination information.
    properties:
      containers:
        description: |-
          Containers is the list of container objects returned for the current page.
          required: true
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerResponse'
        type: array
      metadata:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
        description: |-
          Metadata contains pagination and other metadata for the response.
          required: true
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerResponse:
    description: Contains comprehensive details of a Docker container, including configuration,
      state, and associated resources.
    properties:
      command:
        description: |-
          Command is the command executed when the container started.
          example: "nginx -g daemon off;"
        example: nginx -g daemon off;
        type: string
      container_id:
        description: |-
          ContainerID is the unique identifier assigned by Docker.
          required: true
          example: "f7d9e8c7b6a5..."
        example: f7d9e8c7b6a5
        type: string
      created:
        description: |-
          Created is the timestamp when the container was created by Docker.
          required: true
          example: "2023-10-27T10:00:00Z"
        example: "2023-10-27T10:00:00Z"
        type: string
      created_at:
        description: |-
          CreatedAt is the timestamp when the container record was created in the application database.
          required: true
          example: "2023-10-27T10:00:05Z"
        example: "2023-10-27T10:00:05Z"
        type: string
      finished:
        description: |-
          Finished is the timestamp when the container last finished.
          example: "2023-10-27T11:00:00Z"
        example: "2023-10-27T11:00:00Z"
        type: string
      host_config:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.HostConfigResponse'
        description: HostConfig contains details about the container's host configuration.
      id:
        description: |-
          ID is the internal database ID (if managed by the application).
          example: 101
        example: 101
        type: integer
      image:
        description: |-
          Image is the name of the image used by the container.
          required: true
          example: "nginx:latest"
        example: nginx:latest
        type: string
      image_id:
        description: |-
          ImageID is the ID of the image used by the container.
          required: true
          example: "sha256:a1b2c3d4..."
        example: sha256:a1b2c3d4...
        type: string
      labels:
        additionalProperties:
          type: string
        description: |-
          Labels are the labels applied to the container.
          example: {"environment": "development"}
        example:
          environment: development
        type: object
      name:
        description: |-
          Name is the user-defined name of the container.
          required: true
          example: "my-nginx-container"
        example: my-nginx-container
        type: string
      networks:
        description: Networks lists the network connections for the container.
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkConnectionResponse'
        type: array
      notes:
        description: |-
          Notes are user-defined notes stored in the application database.
          example: "Main web server container."
        example: Main web server container.
        type: string
      platform:
        description: |-
          Platform is the platform string (e.g., "linux/amd64").
          example: "linux/amd64"
        example: linux/amd64
        type: string
      ports:
        description: Ports lists the port mappings for the container.
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.PortMapping'
        type: array
      restart_policy:
        description: |-
          RestartPolicy is the restart policy applied to the container.
          example: "unless-stopped"
        example: unless-stopped
        type: string
      started:
        description: |-
          Started is the timestamp when the container was last started.
          example: "2023-10-27T10:01:00Z"
        example: "2023-10-27T10:01:00Z"
        type: string
      state:
        description: |-
          State is the detailed state string from Docker (e.g., "running", "exited (0)").
          required: true
          example: "running"
        example: running
        type: string
      stats:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerStatsResponse'
        description: Stats contains the latest resource usage statistics (if requested/available).
      status:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerStatus'
        description: |-
          Status is a simplified status string (e.g., running, stopped).
          required: true
          example: "running"
        example: running
      updated_at:
        description: |-
          UpdatedAt is the timestamp when the container record was last updated in the application database.
          required: true
          example: "2023-10-27T10:05:00Z"
        example: "2023-10-27T10:05:00Z"
        type: string
      user_id:
        description: |-
          UserID is the ID of the user who owns/created this container record in the application database.
          required: true
          example: 1
        example: 1
        type: integer
      volumes:
        description: Volumes lists the volume mounts for the container.
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeMountResponse'
        type: array
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.DetailedNetwork:
    properties:
      driver:
        type: string
      id:
        type: string
      internal:
        type: boolean
      ipam_config:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.IPAM'
        description: Use local models.IPAM
      labels:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      name:
        type: string
      options:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      scope:
        type: string
    type: object
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.DetailedVolume:
    properties:
      driver:
        type: string
      labels:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      mountpoint:
        type: string
      name:
        type: string
      options:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
      scope:
        type: string
      status:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.JSONMap'
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorInfo:
    description: Detailed information about an error that occurred during an API request.
    properties:
      code:
        description: |-
          Code is a machine-readable error code identifying the specific error type.
          required: true
          example: RESOURCE_NOT_FOUND
        example: RESOURCE_NOT_FOUND
        type: string
      details:
        description: |-
          Details provides optional additional information about the error, such as validation failures.
          example: {"field": "command", "error": "cannot be empty"}
      message:
        description: |-
          Message is a human-readable description of the error.
          required: true
          example: The requested container was not found.
        example: The requested container was not found.
        type: string
    type: object
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse:
    description: Standard structure for returning errors from the API.
    properties:
      error:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorInfo'
        description: |-
          Error contains the detailed error information.
          required: true
      meta:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
        description: |-
          Meta contains metadata about the response.
          required: true
      success:
        description: |-
          Success indicates if the request was successful (always false for errors).
          required: true
          example: false
        example: false
        type: boolean
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginationResponse'
      request_id:
        example: req-12345
        type: string
      timestamp:
        example: "2023-10-27T10:30:00Z"
        type: string
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.PortMapping:
    description: Defines how a container port is exposed on the host machine.
    properties:
      container_port:
        description: |-
          ContainerPort is the port number inside the container.
          required: true
          example: "80"
        type: string
      host_ip:
        description: |-
          HostIP is the IP address on the host to bind the port to. Defaults to 0.0.0.0 (all interfaces).
          example: "127.0.0.1"
        type: string
      host_port:
        description: |-
          HostPort is the port number on the host. If empty or "0", Docker assigns a random ephemeral port.
          example: "8080"
        type: string
      type:
        description: |-
          Type is the protocol (e.g., "tcp", "udp", "sctp"). Defaults to "tcp".
          example: "tcp"
        type: string
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse:
    properties:
      data:
        description: Use omitempty if data might be nil/empty
      message:
        description: Optional success message
        type: string
      meta:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeMapping:
    description: Defines how a host path or named volume is mounted into a container.
    properties:
      destination:
        description: |-
          Destination is the absolute path inside the container where the volume is mounted.
          required: true
          example: "/var/www/html"
        example: /var/www/html
        type: string
      read_only:
        description: |-
          ReadOnly specifies whether the mount should be read-only within the container.
          example: false
        example: false
        type: boolean
      source:
        description: |-
          Source is the name of the volume or the path on the host machine.
          required: true
          example: "my-app-data" or "/path/on/host"
        example: my-app-data
        type: string
    required:
    - destination
    - source
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Docker Server Manager Go MCP.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Docker Server Manager Go MCP API
  version: "1.0"
paths:
...
  /containers:
    get:
      description: Get a list of containers based on optional filters.
      parameters:
      - default: false
        description: Show all containers (default shows just running)
        in: query
        name: all
        type: boolean
      - default: -1
        description: Maximum number of containers to return
        in: query
        name: limit
        type: integer
      - default: false
        description: Return container sizes
        in: query
        name: size
        type: boolean
      - description: Filters to apply (JSON map format, e.g., {\
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved container list
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerListResponse'
              type: object
        "400":
          description: Invalid filter format
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List containers
      tags:
      - Containers
    post:
      consumes:
      - application/json
      description: Creates a new Docker container based on the provided configuration.
      parameters:
      - description: Container Configuration
        in: body
        name: container
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created container
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.Container'
              type: object
        "400":
          description: Invalid input or configuration
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Permission denied (e.g., creating privileged container as non-admin)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Image not found (if pull fails)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Container name already in use
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error (e.g., Docker daemon error)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new container
      tags:
      - Containers
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"